producer/requirements.txt
kafka-python==1.4.7
robinhood-aiokafka==1.1.6